Cambios servicios

+++ Workspace/Acme-Handy-Worker/src/main/java/services/PhaseService.java
@@ -109,15 +109,20 @@ public class PhaseService {
 
 	public void delete(final Phase phase) {
 		HandyWorker principal;
+		Collection <Application> collApp;
 
 		Assert.notNull(phase);
 		Assert.isTrue(phase.getId() != 0);
 
 		principal = this.handyWorkerService.findByPrincipal();
-
 		Assert.notNull(principal);
-
-		this.phaseRepository.delete(phase);
+		
+		collApp = principal.getApplications();
+		for(Application app : collApp) {
+			if(app.getStatus().equals("ACCEPTED") && app.getFixUpTask().equals(phase.getFixUpTask())){
+				this.phaseRepository.delete(phase);
+			}
+		}
 	}
 
 	// Other business methods
